refactor(group): move`ValidateBasic` logic to `msgServer`
